# project/static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Event Management AlfaBet API",
    "version": "1.0",
    "description": "API documentation"
  },
  "paths": {
    "/register": {
      "post": {
        "summary": "Register a new user",
        "parameters": [
          {
            "name": "User",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/User"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User registered successfully"
          },
          "400": {
            "description": "Email already exists in the database"
          }
        }
      }
    },
    "/login": {
      "post": {
        "summary": "Login and get an authentication token",
        "parameters": [
          {
            "name": "Authorization",
            "in": "header",
            "type": "string",
            "description": "Basic authentication with email and password"
          }
        ],
        "responses": {
          "200": {
            "description": "Token generated successfully"
          },
          "401": {
            "description": "Invalid credentials"
          }
        }
      }
    },
    "/events": {
      "post": {
        "summary": "Schedule a new event",
        "parameters": [
          {
            "name": "Event",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Event"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Event scheduled successfully"
          }
        }
      },
      "get": {
        "summary": "Retrieve a list of all scheduled events",
        "responses": {
          "200": {
            "description": "List of all scheduled events"
          }
        }
      }
    },
    "/events/{event_id}": {
      "get": {
        "summary": "Retrieve details of a specific event",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Details of a specific event"
          },
          "404": {
            "description": "Event not found"
          }
        }
      },
      "put": {
        "summary": "Update details of a specific event",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "type": "integer"
          },
          {
            "name": "Event",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Event"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Event updated successfully"
          },
          "404": {
            "description": "Event not found"
          }
        }
      },
      "delete": {
        "summary": "Delete a specific event",
        "parameters": [
          {
            "name": "event_id",
            "in": "path",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "200": {
            "description": "Event deleted successfully"
          },
          "404": {
            "description": "Event not found"
          }
        }
      }
    }
  },
  "definitions": {
    "User": {
      "type": "object",
      "properties": {
        "email": {
          "type": "string"
        },
        "password": {
          "type": "string"
        }
      }
    },
    "Event": {
      "type": "object",
      "properties": {
        "title": {
          "type": "string"
        },
        "description": {
          "type": "string"
        },
        "venue": {
          "type": "string"
        },
        "event_date": {
          "type": "string"
        },
        "tags": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    }
  }
}